Refactored action facts to base class.
